_performance-golf/pg-creative-os/orion-chief-of-staff/_ops/orion-team-bot/requirements.txt
slack-bolt>=1.18.0
slack-sdk>=3.27.0
anthropic>=0.42.0
python-dotenv>=1.0.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
requests>=2.31.0
